+++ Interstation-Two-WW2/code/modules/reagents/reagent_dispenser.dm
@@ -174,9 +174,7 @@
 		explosion(loc,0,1,3,1)
 	else if (reagents.total_volume > 50)
 		explosion(loc,-1,1,2,1)
-	if(src)
-		spawn (1)
-			qdel(src)
+	if (src) qdel(src)
 
 /obj/structure/reagent_dispensers/fueltank/fire_act(datum/gas_mixture/air, temperature, volume)
 	if (modded)
